+++ Changelog.md
@@ -16,6 +16,7 @@
 ### Internal changes
 * Caused-by lines are now stored together with a taintedness value, which allows filtering taintedness depending on the sink type
 * Taintedness is now stored in a value object, rather than a plain integer.
+* Function taintedness is now stored in a value object, rather than an array of integers.
 
 ## v3.0.4
 ### New features